Added decimal calculations

+++ Calculator Form/Calculator/Calculator/AbsOp.cs
@@ -1,9 +1,9 @@
 ﻿namespace Calculator
 {
     public class AbsOp : UnaryOp {
-        public AbsOp(int operand) : base(operand) {}
+        public AbsOp(double operand) : base(operand) { }
 
-        public override int Evaluate()
+        public override double Evaluate()
         {
             if (_operand < 0) return 0 - _operand;
