+++ Calculator Form/Calculator/Calculator/AddOp.cs
@@ -1,9 +1,9 @@
 ﻿namespace Calculator
 {
     public class AddOp : BinaryOp {
-        public AddOp(int operand1, int operand2) : base(operand1, operand2) {}
+        public AddOp(double operand1, double operand2) : base(operand1, operand2) { }
 
-        public override int Evaluate()
+        public override double Evaluate()
         {
             return Operand1 + Operand2;
         }
