+++ Calculator Form/Calculator/Calculator/BinaryOp.cs
@@ -1,10 +1,11 @@
 ﻿namespace Calculator
 {
     public abstract class BinaryOp : Expression {
-        protected int Operand1;
-        protected int Operand2;
+        protected double Operand1;
+        protected double Operand2;
 
-        protected BinaryOp(int operand1, int operand2) : base() {
+        protected BinaryOp(double operand1, double operand2) : base()
+        {
             Operand1 = operand1;
             Operand2 = operand2;
         }
