+++ Calculator Form/Calculator/Calculator/Expression.cs
@@ -1,6 +1,6 @@
 ﻿namespace Calculator
 {
     public abstract class Expression {
-        public abstract int Evaluate();
+        public abstract double Evaluate();
     };
 }
