+++ Calculator Form/Calculator/Calculator/NegOp.cs
@@ -1,10 +1,10 @@
 ﻿namespace Calculator
 {
     public class NegOp : UnaryOp {
-        public NegOp(int operand) : base(operand) {
+        public NegOp(double operand) : base(operand) {
         }
 
-        public override int Evaluate()
+        public override double Evaluate()
         {
             return -1 * _operand;
         }
