+++ Calculator Form/Calculator/Calculator/SubOp.cs
@@ -2,9 +2,9 @@
 {
     public class SubOp : BinaryOp
     {
-        public SubOp(int operand1, int operand2) : base(operand1, operand2) {}
+        public SubOp(double operand1, double operand2) : base(operand1, operand2) {}
 
-        public override int Evaluate()
+        public override double Evaluate()
         {
             return Operand1 - Operand2;
         }
