+++ Calculator Form/Calculator/Calculator/UnaryOp.cs
@@ -1,9 +1,9 @@
 ﻿namespace Calculator
 {
     public abstract class UnaryOp : Expression {
-        protected int _operand;
+        protected double _operand;
 
-        protected UnaryOp(int operand) {
+        protected UnaryOp(double operand) {
             _operand = operand;
         }
     };
